pfebackend/requirements.txt
asgiref==3.8.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
constantly==23.10.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==44.0.2
daphne==4.1.2
defusedxml==0.7.1
Django==5.1.6
django-cors-headers==4.7.0
django-filter==25.1
django-rest-swagger==2.2.0
django-unfold==0.51.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
dotenv==0.9.9
drf-yasg==1.21.9
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
msgpack==1.1.0
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==24.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==25.0.0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
service-identity==24.2.0
setuptools==75.8.2
simplejson==3.20.1
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
zope.interface==7.2
